refactor: migrate session_manager ProcessError from im::processor to processor_chain
Source: user
Type: refactor
